Use int32 for kExpiredHistogramsHashes instead of int64
The kExpiredHistogramsHashes array is 29kb and grew 11kb in M88. If this
is something that will continue to grow at this rate, we should try and
shrink it.

I used uint32_t for kExpiredHistogramsHashes instead of uint64_t to
reduce array size.

Bug: 1160505
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2821560
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881065}

+++ base/metrics/statistics_recorder.cc
@@ -339,7 +339,7 @@ void StatisticsRecorder::SetRecordChecker(
 }
 
 // static
-bool StatisticsRecorder::ShouldRecordHistogram(uint64_t histogram_hash) {
+bool StatisticsRecorder::ShouldRecordHistogram(uint32_t histogram_hash) {
   const AutoLock auto_lock(lock_.Get());
   EnsureGlobalRecorderWhileLocked();
   return !top_->record_checker_ ||
