+++ components/metrics/expired_histogram_util.cc
@@ -24,7 +24,7 @@ const base::FeatureParam<std::string> kAllowlistParam{
 
 }  // namespace
 
-void EnableExpiryChecker(const uint64_t* expired_histograms_hashes,
+void EnableExpiryChecker(const uint32_t* expired_histograms_hashes,
                          size_t num_expired_histograms) {
   DCHECK(base::FeatureList::GetInstance());
   if (base::FeatureList::IsEnabled(kExpiredHistogramLogicFeature)) {
